requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
notion-client==2.2.1
pydantic==2.5.0
python-dateutil==2.8.2
